split the question-specific UI tests to own test classes
changed the course and instructor ID for each UI test to prevent state leaks

split the question-specific UI tests to own test classes
