+++ src/test/java/teammates/test/cases/ui/browsertests/FeedbackConstSumRecipientQuestionUiTest.java
@@ -1,0 +1,137 @@
+package teammates.test.cases.ui.browsertests;
+
+import static org.testng.AssertJUnit.assertEquals;
+import static org.testng.AssertJUnit.assertNotNull;
+import static org.testng.AssertJUnit.assertNull;
+import static org.testng.AssertJUnit.assertTrue;
+
+import org.testng.annotations.AfterClass;
+import org.testng.annotations.BeforeClass;
+import org.testng.annotations.Test;
+
+import teammates.common.datatransfer.DataBundle;
+import teammates.common.util.Const;
+import teammates.test.driver.BackDoor;
+import teammates.test.pageobjects.Browser;
+import teammates.test.pageobjects.BrowserPool;
+import teammates.test.pageobjects.InstructorFeedbackEditPage;
+
+public class FeedbackConstSumRecipientQuestionUiTest extends FeedbackQuestionUiTest {
+    private static Browser browser;
+    private static InstructorFeedbackEditPage feedbackEditPage;
+    private static DataBundle testData;
+
+    private static String courseId;
+    private static String feedbackSessionName;
+    private static String instructorId;
+    
+    @BeforeClass
+    public void classSetup() throws Exception {
+        printTestClassHeader();
+        testData = loadDataBundle("/FeedbackConstSumRecipientQuestionUiTest.json");
+        removeAndRestoreTestDataOnServer(testData);
+        browser = BrowserPool.getBrowser();
+        
+        instructorId = testData.accounts.get("instructor1").googleId;
+        courseId = testData.courses.get("course").id;
+        feedbackSessionName = testData.feedbackSessions.get("openSession").feedbackSessionName;
+        feedbackEditPage = getFeedbackEditPage(instructorId, courseId, feedbackSessionName, browser);
+
+    }
+    
+    @Test
+    public void allTests() throws Exception{
+        testEditPage();
+        
+        //TODO: move/create other MCQ question related UI tests here.
+        //i.e. results page, submit page.
+    }
+    
+    private void testEditPage(){
+        testNewQuestionFrame();
+        testInputValidation();
+        testCustomizeOptions();
+        testAddQuestionAction();
+        testEditQuestionAction();
+        testDeleteQuestionAction();
+    }
+
+    public void testNewQuestionFrame() {
+        ______TS("CONSTSUM-recipient: new question (frame) link");
+
+        feedbackEditPage.selectNewQuestionType("Distribute points (among recipients) question");
+        feedbackEditPage.clickNewQuestionButton();
+        assertTrue(feedbackEditPage.verifyNewConstSumQuestionFormIsDisplayed());
+    }
+    
+    public void testInputValidation() {
+        
+        ______TS("CONST SUM:input validation");
+        
+        feedbackEditPage.fillQuestionBox("ConstSum-recipient qn");
+        feedbackEditPage.fillConstSumPointsBox("", -1);
+        
+        assertEquals("1", feedbackEditPage.getConstSumPointsBox(-1));
+        assertEquals(false, feedbackEditPage.isElementVisible("constSumOptionTable--1"));
+        
+        feedbackEditPage.getDeleteQuestionLink(-1).click();
+        assertEquals("", feedbackEditPage.getStatus());
+        assertNull(BackDoor.getFeedbackQuestion(courseId, feedbackSessionName, 1));
+        
+    }
+
+    public void testCustomizeOptions() {
+        feedbackEditPage.selectNewQuestionType("Distribute points (among recipients) question");
+        feedbackEditPage.clickNewQuestionButton();
+        
+        ______TS("CONST SUM: set points options");
+
+        feedbackEditPage.selectConstSumPointsOptions("per recipient:", -1);
+        feedbackEditPage.fillConstSumPointsBox("100", -1);
+        
+    }
+
+    public void testAddQuestionAction() {
+        ______TS("CONST SUM: add question action success");
+        
+        feedbackEditPage.fillQuestionBox("const sum qn");
+        feedbackEditPage.selectRecipientsToBeStudents();
+        assertNull(BackDoor.getFeedbackQuestion(courseId, feedbackSessionName, 1));
+        feedbackEditPage.clickAddQuestionButton();
+        assertEquals(Const.StatusMessages.FEEDBACK_QUESTION_ADDED, feedbackEditPage.getStatus());
+        assertNotNull(BackDoor.getFeedbackQuestion(courseId, feedbackSessionName, 1));
+        feedbackEditPage.verifyHtmlMainContent("/instructorFeedbackConstSumRecipientQuestionAddSuccess.html");
+    }
+
+    public void testEditQuestionAction() {
+        ______TS("CONST SUM: edit question success");
+
+        assertEquals(true, feedbackEditPage.clickEditQuestionButton(1));
+        feedbackEditPage.fillEditQuestionBox("edited const sum qn text", 1);
+        feedbackEditPage.fillConstSumPointsBox("200", 1);
+        feedbackEditPage.selectConstSumPointsOptions("in total:", 1);
+        
+        feedbackEditPage.clickSaveExistingQuestionButton(1);
+        assertEquals(Const.StatusMessages.FEEDBACK_QUESTION_EDITED, feedbackEditPage.getStatus());
+
+        feedbackEditPage.verifyHtmlMainContent("/instructorFeedbackConstSumRecipientQuestionEditSuccess.html");
+    }
+    
+    public void testDeleteQuestionAction(){
+        ______TS("CONSTSUM: qn delete then cancel");
+
+        feedbackEditPage.clickAndCancel(feedbackEditPage.getDeleteQuestionLink(1));
+        assertNotNull(BackDoor.getFeedbackQuestion(courseId, feedbackSessionName, 1));
+
+        ______TS("CONSTSUM: qn delete then accept");
+
+        feedbackEditPage.clickAndConfirm(feedbackEditPage.getDeleteQuestionLink(1));
+        assertEquals(Const.StatusMessages.FEEDBACK_QUESTION_DELETED, feedbackEditPage.getStatus());
+        assertNull(BackDoor.getFeedbackQuestion(courseId, feedbackSessionName, 1));    
+    }
+    
+    @AfterClass
+    public static void classTearDown() throws Exception {
+        BrowserPool.release(browser);
+    }
+}
